scripting: add sha to script cache after SCRIPT LOAD
Propagate the script only when it isn't in script cache,
thus we can save our IO and bandwidth.

And then add it into the script cache, as from now on
slaves and AOF know about it.

Moreover, if the script is read-only, after this when
we execute EVALSHA, we don't need to rewrite EVALSHA
as EVAL and propagate the script.