scripting & replication: only persist scripts in cache to rdb
On master, we only persist scripts in cache instead all scripts,
because what want slave to have is the scripts we have already
replicated. And we do not flush script cache when full resync.

On slave, after full resync slave should reset the script state,
including server.lua_scripts and script cache.

And when we load Lua scripts from rdb, we also need to add them
into script cache.

The key point is, for lua scripts, what we want our slave to have
is what we have propagated, instead of what we have.