aof: update aof_last_fsync time only when we really do bio fsync
The sync_in_progress background job is not what we want,
it's the last fsync, can't make sure if it can fsync the
new written data.

So, do not update aof_last_fsync time when we have fsync
in progress, then we can do real bio fsync in the next
flushAppendOnlyFile.